save/load for most of all classes reimplemented
Some classes were cut off. Progect works, but falls with exeption.

+++ Turret_1/code/game_interface/settings/settings/display_settings.h
@@ -21,8 +21,8 @@ struct DisplaySettings
 	{
 		const auto table = root->get_table("display");
 		//
-		windowMaxSize.x = table->get_as<double>("window-max-size-x").value_or(1920);
-		windowMaxSize.y = table->get_as<double>("window-max-size-y").value_or(1016);
+		windowMaxSize.x = table->get_as<uint32_t>("window-max-size-x").value_or(1920);
+		windowMaxSize.y = table->get_as<uint32_t>("window-max-size-y").value_or(1016);
 	}
 
 };
